chore: Update file structre to separate auth and scraper endpoints into separate modules
